refactor : separate get date function
